+++ SplinterlandsRObot/API/Bot.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using SplinterlandsRObot.Constructors;
+using SplinterlandsRObot.Models;
 using SplinterlandsRObot.Extensions;
 using SplinterlandsRObot.Net;
